Add log-tailing containers for extra mon logs
Use 'minimal' image (defaulting to Alpine) to add sidecar
containers to daemonset-mon, along with a small shell script,
to tail the 'extra' monitor logs (ceph.log and ceph.audit.log),
and watch them until they get greater than a certain size
(default 100MB), at which point rename them (to retain one
level of backup log inside the container).  This is simplistic
log rotation because the primary logging happens through Kubernetes.
Also, the daemon isn't signaled to reopen the logs because
it opens the files exclusively for each log message (unlike the
main log, which is assumed to be coming to stderr from the
primary ceph-mon container).

Signed-off-by: Dan Mick <dmick@redhat.com>